requirements.txt
fastapi
gunicorn
uvicorn[standard]
Jinja2
typer
pydantic
httpx
aioredis
pyjwt
rust_decimal
